bmt, pot: moved into swarm package
