+++ swarm/pss/client/doc.go
@@ -15,7 +15,7 @@
 //  	pss "github.com/ethereum/go-ethereum/swarm/pss/client"
 //  	"github.com/ethereum/go-ethereum/p2p/protocols"
 //  	"github.com/ethereum/go-ethereum/p2p"
-//  	"github.com/ethereum/go-ethereum/pot"
+//  	"github.com/ethereum/go-ethereum/swarm/pot"
 //  	"github.com/ethereum/go-ethereum/log"
 //  )
 //
